02/26
The front-end and back-end communication is setup from the back-end point. The server is listening on port 1337, which is where the front-end will have to direct the requests.

The protocol is not TLS encrypted, and I need to read more into setting it up with Angular, but for now it's too much work.

Starting both the server and the app is done by running command npm start. I've modified the package.json file to be able to run on any OS.

Codegangsta/gin is not used to start the server because live updates are not necessary for now.