04/18 Made changes to how courses are stored
The courses and assignments are now stored in two places. The databases for those two contain all contextual information such as the names and codes for courses, and names, descriptions, points, etc. for assignments. The user model now stores IDs of assignments the user has, ID of the course they are enrolled in (this can now easily be updated to store information about more than one course), which are used to query the databases for desired information. Getter functions are also added to send the information about an assignment to the front-end. Also, two accounts (AOCTest and AOCAdmin) are now automatically added into the database if they are not there already. AOCTest can be used for functionality testing, and its password is hardcoded in the back_end/src/helper/helper.go file. This account has 5 achievements, and a couple of assignments, so that can be used to showcase the achievements page.